Distribute DateToEpoch components to smaller ones
This segregated DateToEpoch components to smaller components, make the code easier to read. Created a types folder to store types.
