+++ src/types/timestamp.ts
@@ -1,0 +1,5 @@
+export type Timestamp = {
+  hour: number;
+  minute: number;
+  second: number;
+};
